# SmartHelpDeskSwagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Helpdesk CRM — Swagger đầy đủ",
    "version": "1.0.0",
    "description": "Đối chiếu source ngày 2026-09-12. 25 handler nghiệp vụ / 29 method-path (4 alias auth). JSON UTF-8; UUID dạng chuỗi; Instant ISO-8601 UTC. ResponseGeneral bỏ field null ở cấp envelope. HTTP 201 vẫn có body.status=200. Lỗi Spring Security filter không bảo đảm envelope ApiError. AI chạy RuleBasedLlmAdapter nội bộ, không gọi API LLM bên ngoài."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local; đổi host/port theo môi trường"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/requests": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Tạo yêu cầu hỗ trợ",
        "operationId": "createRequest",
        "description": "Quyền: CLIENT. Client phải có companyId. Tạo PENDING, chưa gán developer, ghi CREATE history và gửi alert cho admin (HIGH_PRIORITY_REGISTERED nếu HIGH, còn lại REQUEST_REGISTERED).",
        "x-access": "CLIENT",
        "responses": {
          "201": {
            "description": "Tạo thành công; body.status=200.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createRequestResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Request created successfully",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "title": "Không đăng nhập được",
                    "description": "Ví dụ",
                    "category": "BUG",
                    "priority": "HIGH",
                    "status": "PENDING",
                    "clientId": "0198f101-1234-7000-8000-000000000001",
                    "clientName": "Ví dụ",
                    "assignedDeveloperId": "0198f101-1234-7000-8000-000000000001",
                    "assignedDeveloperName": "Ví dụ",
                    "createdAt": "2026-09-12T03:00:00Z",
                    "updatedAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              },
              "example": {
                "title": "Không đăng nhập được",
                "description": "Ứng dụng báo error khi đăng nhập.",
                "category": "BUG",
                "priority": "MEDIUM"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Danh sách yêu cầu",
        "operationId": "listRequests",
        "description": "Quyền: Đã đăng nhập. ADMIN xem tất cả; CLIENT chỉ request mình tạo; DEVELOPER chỉ request được giao. Bộ lọc kết hợp AND. search tìm title/description không phân biệt hoa thường, trim khoảng trắng; % và _ giữ wildcard SQL LIKE. page âm, size <= 0, sortBy không hợp lệ hoặc lỗi chuyển kiểu query/path có thể vào SYS_500.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/listRequestsResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "content": [],
                    "pageable": {
                      "pageNumber": 0,
                      "pageSize": 0,
                      "sort": {
                        "empty": false,
                        "sorted": false,
                        "unsorted": false
                      },
                      "offset": 0,
                      "paged": false,
                      "unpaged": false
                    },
                    "totalPages": 0,
                    "totalElements": 0,
                    "last": false,
                    "size": 0,
                    "number": 0,
                    "sort": {
                      "empty": false,
                      "sorted": false,
                      "unsorted": false
                    },
                    "numberOfElements": 0,
                    "first": false,
                    "empty": false
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RequestCategory"
            },
            "description": "Lọc loại yêu cầu"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RequestPriority"
            },
            "description": "Lọc ưu tiên"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RequestStatus"
            },
            "description": "Lọc trạng thái"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Tìm title hoặc description"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            },
            "description": "Trang bắt đầu từ 0"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1
            },
            "description": "Số bản ghi/trang; chưa giới hạn tối đa"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "description": "Thuộc tính entity Request: createdAt, updatedAt, title, priority, status..."
          },
          {
            "name": "sortDir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc"
            },
            "description": "asc không phân biệt hoa thường là tăng; mọi giá trị khác dùng giảm"
          }
        ]
      }
    },
    "/api/requests/stats": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Thống kê yêu cầu",
        "operationId": "requestStats",
        "description": "Quyền: ADMIN. Tổng số, số DONE, tỷ lệ hoàn thành, số theo category và developer chưa xóa.",
        "x-access": "ADMIN",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/requestStatsResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "totalRequests": 0,
                    "completedRequests": 0,
                    "completionRate": 0,
                    "requestsByCategory": {},
                    "requestsByDeveloper": []
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/requests/{id}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Chi tiết yêu cầu",
        "operationId": "requestDetail",
        "description": "Quyền: ADMIN hoặc client tạo request hoặc developer được giao. Gồm công ty, client, developer và history tăng dần theo changedAt. Sai quyền tài nguyên trả BAD_400.",
        "x-access": "ADMIN hoặc client tạo request hoặc developer được giao",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/requestDetailResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "companyAddress": "Ví dụ",
                    "companyPhone": "Ví dụ",
                    "title": "Không đăng nhập được",
                    "description": "Ví dụ",
                    "category": "BUG",
                    "priority": "HIGH",
                    "status": "PENDING",
                    "clientId": "0198f101-1234-7000-8000-000000000001",
                    "clientName": "Ví dụ",
                    "clientEmail": "Ví dụ",
                    "assignedDeveloperId": "0198f101-1234-7000-8000-000000000001",
                    "assignedDeveloperName": "Ví dụ",
                    "assignedDeveloperEmail": "Ví dụ",
                    "createdAt": "2026-09-12T03:00:00Z",
                    "updatedAt": "2026-09-12T03:00:00Z",
                    "history": []
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      }
    },
    "/api/requests/{id}/assign": {
      "patch": {
        "tags": [
          "Requests"
        ],
        "summary": "Phân công developer",
        "operationId": "assignDeveloper",
        "description": "Quyền: ADMIN. Không body, {} hoặc developerId=null: chọn developer active, chưa xóa, ít task PENDING/IN_PROGRESS nhất; hòa thì chọn người hoàn thành gần nhất, tiếp tục hòa chọn ID nhỏ nhất theo chuỗi. Gán thủ công chỉ kiểm tra tồn tại và role DEVELOPER. Không đổi status; ghi ASSIGN và gửi ASSIGNED alert.",
        "x-access": "ADMIN",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/assignDeveloperResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Developer assigned successfully",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "title": "Không đăng nhập được",
                    "description": "Ví dụ",
                    "category": "BUG",
                    "priority": "HIGH",
                    "status": "PENDING",
                    "clientId": "0198f101-1234-7000-8000-000000000001",
                    "clientName": "Ví dụ",
                    "assignedDeveloperId": "0198f101-1234-7000-8000-000000000001",
                    "assignedDeveloperName": "Ví dụ",
                    "createdAt": "2026-09-12T03:00:00Z",
                    "updatedAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDeveloperDto"
              },
              "examples": {
                "manual": {
                  "value": {
                    "developerId": "0198f101-1234-7000-8000-000000000001"
                  }
                },
                "automatic": {
                  "value": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/requests/{id}/status": {
      "patch": {
        "tags": [
          "Requests"
        ],
        "summary": "Cập nhật trạng thái",
        "operationId": "updateStatus",
        "description": "Quyền: ADMIN hoặc developer được giao. Chỉ PENDING → IN_PROGRESS → DONE. Không cho cùng trạng thái, đi lùi, bỏ bước hoặc mở lại DONE. Ghi UPDATE/CLOSE; gửi STATUS_CHANGED cho client. Sai quyền trả BAD_400.",
        "x-access": "ADMIN hoặc developer được giao",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateStatusResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Request status updated successfully",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "title": "Không đăng nhập được",
                    "description": "Ví dụ",
                    "category": "BUG",
                    "priority": "HIGH",
                    "status": "PENDING",
                    "clientId": "0198f101-1234-7000-8000-000000000001",
                    "clientName": "Ví dụ",
                    "assignedDeveloperId": "0198f101-1234-7000-8000-000000000001",
                    "assignedDeveloperName": "Ví dụ",
                    "createdAt": "2026-09-12T03:00:00Z",
                    "updatedAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestStatusDto"
              },
              "example": {
                "status": "IN_PROGRESS",
                "memo": "Bắt đầu xử lý"
              }
            }
          }
        }
      }
    },
    "/api/requests/{id}/history": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Lịch sử yêu cầu",
        "operationId": "requestHistory",
        "description": "Quyền: ADMIN hoặc client tạo request hoặc developer được giao. changedAt tăng dần. REOPEN có trong enum nhưng không có luồng mở lại.",
        "x-access": "ADMIN hoặc client tạo request hoặc developer được giao",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/requestHistoryResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": []
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      }
    },
    "/api/requests/classify": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Phân loại yêu cầu",
        "operationId": "classify",
        "description": "Quyền: Đã đăng nhập. Rule-based: keyword BUG (.92), rồi FEATURE (.88), còn lại INQUIRY (.80). Không gọi nhà cung cấp AI.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/classifyResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "category": "BUG",
                    "confidence": 0.92,
                    "reason": "Matched defect/error keyword: 'error'"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequestDto"
              },
              "example": {
                "description": "Application error when logging in"
              }
            }
          }
        }
      }
    },
    "/api/requests/suggest-priority": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Gợi ý độ ưu tiên",
        "operationId": "suggestPriority",
        "description": "Quyền: Đã đăng nhập. Rule-based: keyword HIGH (.95), rồi LOW (.85), còn lại MEDIUM (.80).",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/suggestPriorityResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "priority": "HIGH",
                    "confidence": 0.95,
                    "reason": "Matched critical/urgency keyword: 'production'"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestPriorityDto"
              },
              "example": {
                "description": "Production is unavailable"
              }
            }
          }
        }
      }
    },
    "/api/requests/{id}/summary": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "Tóm tắt yêu cầu",
        "operationId": "requestSummary",
        "description": "Quyền: Đã đăng nhập. Hiện chỉ kiểm tra request tồn tại, không kiểm tra quyền sở hữu/phân công. Mẫu: [category] title - Request status is status with priority priority.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/requestSummaryResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "requestId": "0198f101-1234-7000-8000-000000000001",
                    "summary": "[BUG] Không đăng nhập được - Request status is PENDING with MEDIUM priority."
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Dán accessToken từ login; Authorization: Bearer <accessToken>."
      }
    },
    "schemas": {
...
      "RequestCategory": {
        "type": "string",
        "enum": [
          "BUG",
          "FEATURE",
          "INQUIRY"
        ]
      },
...
      "createRequestResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Request created successfully"
          },
          "data": {
            "$ref": "#/components/schemas/RequestResponse"
          }
        }
      },
...
    },
    "responses": {
      "Error400": {
        "description": "VAL_400: body/validation sai. BAD_400: nghiệp vụ sai hoặc không có quyền tài nguyên.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 400,
                  "errorCode": "BAD_400",
                  "message": "Bad request"
                }
              },
              "validation": {
                "value": {
                  "status": 400,
                  "errorCode": "VAL_400",
                  "message": "Email cannot be blank",
                  "data": {
                    "email": "Email cannot be blank"
                  }
                }
              }
            }
          }
        }
      },
...
      "Error404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 404,
                  "errorCode": "RES_404",
                  "message": "Resource not found"
                }
              }
            }
          }
        }
      },
      "Error500": {
        "description": "System internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 500,
                  "errorCode": "SYS_500",
                  "message": "System internal error"
                }
              }
            }
          }
        }
      },
      "SecurityDenied": {
        "description": "403: thiếu/token không hợp lệ hoặc không đủ role ở security filter; body do framework quyết định, có thể rỗng. AccessDeniedException đến controller advice trả ApiError với AUTH_403."
      }
    }
  }
}